==> code/2 - Instance Segmentation.ipynb <==
# Dependencies
import numpy as np
from PIL import Image
import open3d as o3d
import json
import os
from tqdm import tqdm
import time
import pandas as pd

from sklearn.cluster import DBSCAN

import matplotlib.pyplot as plt
from matplotlib.axes import Axes
from matplotlib import rcParams

import os.path as osp
from nuscenes import NuScenes

# Utils for Lidar and Radar
from nuscenes.nuscenes import NuScenes
from utils.misc import *
from collections import defaultdict

%matplotlib inline

#establishing root data and folder for nuscenes
nuscenes_root = "/users/CeliaSagastume/Desktop/Lidar-Net/data/sets/nuscenes"
nusc = NuScenes(version='v1.0-mini', dataroot=nuscenes_root, verbose=True)

#Function to Segment all point clouds in the dataset and save them with corresponding object label

def Save_PointClouds(FilePath):
    
    #define samples path
    samples_path = FilePath

    # define path to save point clouds in .csv
    csv_path =  os.getcwd() + '/dataset/data/'

    #Dictionary to store point clouds 
    pc_dict = defaultdict( lambda: defaultdict(list))

    #Dictionary to store categories
    cat_dict = defaultdict(list)

    # extract all files in path
    for root, dirs, files in os.walk(samples_path):

        # select each sample file
        for file_ in tqdm(files, desc = "Sample files"):   

            # load the current sample file
            data_json = load_file(samples_path + file_)

            # process each instance
            for index in range(len(data_json['instance'])):

                # Create instance metadata
                annotation_metadata = data_json['instance'][index]

                #Retrieve instance token 
                category_name = data_json['instance'][index]['category']

                #Retrieve annotation token
                annotation = annotation_metadata['annotation_token']

                # point cloud
                pointcloud_lidar = load_pcl_txt(annotation_metadata['pointcloud_path'], annotation_metadata['pcl_shape'])


                clustering = DBSCAN(eps = 1.75, min_samples = 20).fit(pointcloud_lidar.T)
                # Clusters
                dbscan_clusters = clustering.labels_

                # Obtain the more repeated index different of -1
                class_idx = np.bincount(dbscan_clusters[dbscan_clusters != -1]).argmax()
                # Find indexes of interest cluster
                indexes = np.where(dbscan_clusters == class_idx)[0]

                # extract the bigger cluster
                pcl_temp = pointcloud_lidar.T[indexes]
                
                #create point cloud array
                Velopoints = np.asarray(pcl_temp, np.float32)

                   #add point cloud to dictionary 
                pc_dict[annotation]['pointcloud'].append(Velopoints.tolist())

                #add category label to category dictionary
                pc_dict[annotation]['category'].append(category_name)

            #convert dictionary to json file and save for modelling 
            with open('pc_data.txt', 'w') as fp:
                json.dump(pc_dict, fp,  indent=4)


            

Path = os.getcwd() + '/dataset/samples/'
Save_PointClouds(Path)

Dataset and Instance segmentation adapted from:

@misc{salazargomez2021highlevel,
      title={High-level camera-LiDAR fusion for 3D object detection with machine learning},
      author={Gustavo A. Salazar-Gomez and Miguel A. Saavedra-Ruiz and Victor A. Romero-Cano},
      year={2021},
      eprint={2105.11060},
      archivePrefix={arXiv},
      primaryClass={cs.CV}
}
